Release 0.0.6 - we now initialize the library lazily
Rather than requre the user to call init() we do it for them.
This is done the first time a routine is called rather than when
the module is loaded to allow users to import this module without
a working XSPEC installation (although, to be honest, it's probably
not worth the complexity as I'm not sure you can load it without
a working XSPEC installation).

The initalization step normally displays the default abundance
and cross-section tables. We hide this output.