Release 0.0.15 - convolution models are now in place; other models optional
The convolution models now over-write the contents of the model
argument (they still return the convolved model). Other models
can select this behavior by sending in the NumPy array with the
out parameter - for example

    y = np.ones(egrid.size - 1)
    x.apec(energies=egrid, pars=pars, out=y)

Fixes #7 - it's not clear whether it saves any time (I assume that
all the overhead of calling the correct code and format conversions
mean that the cost of allocating memory is normally irrelevant).