Release 0.0.17 - add an explicit Python module
The compiled code now generated a _compiled module which is then
re-exported by the xspec_models_cxc/__init__.py module - so there
should be no difference for a hypothetical user.

This is intended to support adding model-specific information to
the module but it does not do that yet. I am not 100% sure that
the build setup is correct.