Release 0.0.22 - tableModel now has optional model argument
In working on this I've realised the current approach to this
"work in place" setup is not ideal since if you don't send in
the correct argument type (e.g. np.float32 for tableModel) then
the array will be copied and so not updated in place. It should
really error out.