Version 0.0.26: bump to XSPEC 12.12.1
a) update the setup code to try and be less version specific,
   in particular for the library names we require to link to
   (i.e. allow libwcs-*.so rather than requiring a particular
   value)

b) allow support for XSPEC 12.12.1; unfortunately this means
   dropping 12.12.0 support

Unfortunately the table model code has changed between the two
and it's not obvious to me how we can deal with this at build
time, as we aren't able to identify the XSPEC version in play.

The problem here is that tabint is not in any XSPEC 12.12.0
include file, so we need to add a declaration. The symbol is
available in 12.12.1, but it has changed (a bunch of arguments
are now labelled as const), which conflicts with the 12.12.0
version. In Sherpa we handle this by explicitly requiring a
version label from the user.

There is also the issue that in 12.12.1 we need to support
Fortran double precision models. I have added code to support
them, but it is untested. It also replicates the single-precision
code, and I'm sure that a C++ template whiz could remove the
duplication.