Version 0.0.27 - get the XSPEC version at build time
We can get the XSPEC version number by building a program
which calls the necessary XSPEC "get me the version" routine.
This is definitely OTT but I'm interested to see how useful it
turns out to be.

At the moment we don't really take advantage of this information,
and the build is hard-coded to prefer g++ rather than clang.