Update documnetation for XSPEC 12.14.1
The code builds for XSPEC 12.14.1 (and so presumably 12.14.0)
so update the documentation. As part of this, remove some of the
references to older XSPEC versions as they are not useful.

Add scripts/report_models.py to auto-generate the text needed for
the README. The code repeats logic in helpres/template.py which
could be absracted out (maybe included in the '__init__.py' file
directly), but leave that for a later revision.