Support building against CXC xspec-modelsonly conda package
Allow for building against the CXC xspec-modelsonly conda package,
which is fun because $HEADAS does not contain lib/ and include/
but $HEADAS/../ does.

As part of this allow the code we build to get the current XSPEC
version to hard-code the library paths in it.

The setuptools configuration has been updated to use pyproject.toml
for configuration, although it's still not completely declarative.

The code will hopefully now build on macOS systems.

Files are not necessarily installed in sensible locations.