Rework the build logic
The build logic had got a bit confused, so separate out some
of the functionality. There is more that could be done, but
let's take some improvements.

The main aim was to separate out the logic steps needed and
make them separate scripts in helpers/, and then to actually
call them rather than import them, in setup.py. This is a bit
annoying, as it's hard to send around complex state, but the
new version should make it a bit clearer what functionality
is needed.

The aim was to move from setuptools to meson, but I decided
that there was enough to clean up to make it worth doing this
step first.