Rework how the template code is organized
We now provide the code used to build the interface - both the
compiled/C++ and Python code - as part of the Python module. The
reason for this is to support XSPEC user models (although this
work has not been finalized yet).

This is complicated because we can't easily use the module to
build the module, and we don't want to repeat the logic in
multiple places (as there's a chance that the code will diverge).
The approach taken is to have the logic in a file in helpers/
that is used

a) to build the code
b) is copied into the Python module

As part of this we require parse-xspec to be installed when
testing the module. It is also needed if you ever want to use
XSPEC user models, but this has not been worked out yet so
we do not make it a requirement for general use.